Map tab movement instead of window movement
For the longest time I had S-J remapped, but this is used to join lines
in Vim which is quite useful. So I've found there's not a great way to
map both window and tab movement keys at the same time. But, the default
window movement keybindings in Vim aren't that bad. So rather than make
new mappings for window movement only slightly better than the default,
I'll use the default keybindings for window movement and use the
Shift-movement keys for tabs.